Added few modications

+++ app/api_v1/orders.py
@@ -1,10 +1,11 @@
 from flask import request
 from flask_login import login_required, current_user
+from datetime import datetime
 from . import v1_api as api
 from .. import db
 from ..models import Order
 from ..decorators import json, paginate
-from ..utils import admin_required
+from ..utils import admin_required, send_error, log_activity
 
 
 @api.route('/orders/', methods=['GET'])
@@ -15,7 +16,7 @@ def get_orders():
     return db.session.query(Order).filter_by(company_id=current_user.company_id).all()
 
 
-@api.route('/orders/<int:id>', methods=['GET'])
+@api.route('/orders/<int:order_id>', methods=['GET'])
 @json
 @admin_required
 def get_customer_orders(order_id):
@@ -27,29 +28,35 @@ def get_customer_orders(order_id):
 @login_required
 @json
 def new_customer_order():
+    order_data = request.get_json()
+    if order_data is None:
+        return send_error(400, 'Bad request', 'This request contains invalid or no data')
+    payment_id, order_object = Order.import_data(order_data)
+    if payment_id is None or order_object is None:
+        return send_error(400, 'Bad request', 'Missing data in order form')
+    new_order = Order(staff_id=current_user.id, date_of_order=datetime.now(),
+                      payment_reference=payment_id,
+                      company_id=current_user.company_id, items=order_object)
+    try:
+        db.session.add(new_order)
+        db.session.commit()
+        return {}, 201, {'Message': 'Successful'}
+    except Exception as e:
+        log_activity('EXCEPTION[new_customer_order]', current_user.username,'',
+                     str(e))
+        return send_error(404, 'Bad request', 'There was an error processing '
+                                              'the data sent in your request')
 
-    customer = Customer.query.get_or_404(id)
-    order = Order(customer=customer)
-    order.import_data(request.json)
-    db.session.add(order)
-    db.session.commit()
-    return {}, 201, {'Location': order.get_url()}
 
-
-@api.route('/orders/<int:id>', methods=['PUT'])
+@api.route('/orders/<int:order_id>', methods=['DELETE'])
+@admin_required
 @json
-def edit_order(id):
-    order = Order.query.get_or_404(id)
-    order.import_data(request.json)
-    db.session.add(order)
-    db.session.commit()
-    return {}
-
-
-@api.route('/orders/<int:id>', methods=['DELETE'])
-@json
-def delete_order(id):
-    order = Order.query.get_or_404(id)
-    db.session.delete(order)
-    db.session.commit()
-    return {}
+def delete_order(order_id):
+    order = db.session.query(Order).filter_by(company_id=current_user.company_id,
+                                              id=order_id).first()
+    if order is not None:
+        db.session.delete(order)
+        db.session.commit()
+        return {}
+    return send_error(404, 'Bad request',
+                      'There was an error processing the data sent in your request')
